refactor : 로또 금액을 구하는 부분에서 금액을 통해 로또 갯수를 구하도록 변경
